fix: docker.sh could not rm a stopped container
We make use of 'docker ps's filtering and quiet output: it will only print
the ID of the container if there is a container that starts with $container_name

+++ scripts/docker.sh
@@ -68,7 +68,7 @@ run()
 
 rm()
 {
-    docker kill "$container_name"
+    [ "$(docker ps -q -f "name=$container_name")" ] && docker kill "$container_name"
     docker container rm "$container_name"
 }
